Pass spark task attempt number to TaskId
Instead of passing the default value of 0 to TaskId inplace of stage execution
attempt number, we are passing task attempt number. This is an interim
solution while a long term solution is being looked into. This also resolves a
problem of same taskId being re-used in presto-native result in data
corruption.